biocup_app/back/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
motor
passlib[bcrypt]
python-jose[cryptography]
qdrant-client